docs: salvage capability docs improvements from #4832; promote `TemplateStr` to public `pydantic_ai.template`
Salvaged from dsfaccini's closed PR #4832, adapted to the new capabilities
docs layout and verified against current code:

- Promote `_template.py` -> public `pydantic_ai.template` module with its
  own API reference page (`pydantic_ai.TemplateStr` re-export unchanged;
  internal from_spec helpers stay in `_template.py`)
- Fix `on_tool_validate_error` signature in the hooks table: the error is
  `ValidationError | ModelRetry`, not `Exception`
- Link every type in the lifecycle-hook signature tables
- Document error-hook chaining (reverse capability order, first result
  recovers, re-raise feeds the next handler) and skip-exception
  short-circuit behavior
- Document `for_run` re-extraction semantics (config methods re-run on the
  replacement instance; `get_model`/bootstrap `resolve_model_id` excepted)
  in the docs and the `AbstractCapability` docstring, traced against
  `agent/__init__.py` run setup
- Add a per-capability local-fallback table to the provider-adaptive
  section, a 'Typing dependencies' section with a typed-deps example, and
  clarified hook-ordering, state-sharing, and testing notes
- Drop stray backslashes that rendered literally in `| None` table spans
