+++ docs/api/template.md
@@ -1,0 +1,3 @@
+# `pydantic_ai.template`
+
+::: pydantic_ai.template
